Moved natural sort functions from wxString to wxArrayString.
