+++ include/wx/arrstr.h
@@ -47,6 +47,12 @@ wxDictionaryStringSortDescending(const wxString& s1, const wxString& s2)
     return wxDictionaryStringSortAscending(s2, s1);
 }
 
+int wxCMPFUNC_CONV wxNaturalStringSortAscending(const wxString& s1, const wxString& s2);
+int wxCMPFUNC_CONV wxNaturalStringSortDescending(const wxString& s1, const wxString& s2);
+int wxCMPFUNC_CONV wxNaturalStringSortNoCaseAscending(const wxString& s1, const wxString& s2);
+int wxCMPFUNC_CONV wxNaturalStringSortNoCaseDescending(const wxString& s1, const wxString& s2);
+
+
 #if wxUSE_STD_CONTAINERS
 
 #include "wx/dynarray.h"
